Separate QEMU and CH implementation to different files
