+++ pkg/hypervisor/BUILD.bazel
@@ -2,7 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["hypervisor.go"],
+    srcs = [
+        "ch-hypervisor.go",
+        "hypervisor.go",
+        "qemu-hypervisor.go",
+    ],
     importpath = "kubevirt.io/kubevirt/pkg/hypervisor",
     visibility = ["//visibility:public"],
 )
