+++ pkg/hypervisor/hypervisor.go
@@ -2,10 +2,6 @@ package hypervisor
 
 import "regexp"
 
-// TODO These global variables should be changed to accessor functions in the Hypervisor interface
-var HypervisorDaemonExecutables []string = []string{"virtqemud", "virtchd"}
-var QemuProcessExecutablePrefixes []string = []string{"qemu-system", "qemu-kvm", "cloud-hypervisor"}
-
 // Hypervisor interface defines functions needed to tune the virt-launcher pod spec and the libvirt domain XML for a specific hypervisor
 type Hypervisor interface {
 	// The `ps` RSS for virt-launcher-monitor
@@ -51,156 +47,6 @@ type Hypervisor interface {
 
 	// Return the domain type in Libvirt domain XML
 	GetDomainType() string
-}
-
-// Define QemuHypervisor struct that implements the Hypervisor interface
-type QemuHypervisor struct {
-}
-
-type CloudHypervisor struct {
-}
-
-// Implement GetDomainType method for QemuHypervisor
-func (q *QemuHypervisor) GetDomainType() string {
-	return "kvm"
-}
-
-// Implement SupportsMemoryBallooning method for QemuHypervisor
-func (q *QemuHypervisor) SupportsMemoryBallooning() bool {
-	return true
-}
-
-// Implement RequiresBoot order method for QemuHypervisor
-func (q *QemuHypervisor) RequiresBootOrder() bool {
-	return false
-}
-
-// Implement GetDiskDriver method for QemuHypervisor
-func (q *QemuHypervisor) GetDiskDriver() string {
-	return "qemu"
-}
-
-// Implement GetLibvirtSocketPath method for QemuHypervisor
-func (q *QemuHypervisor) GetLibvirtSocketPath() string {
-	return "libvirt/libvirt-sock"
-}
-
-// Implement GetVcpuRegex method for QemuHypervisor
-func (q *QemuHypervisor) GetVcpuRegex() *regexp.Regexp {
-	// parse thread comm value expression
-	return regexp.MustCompile(`^CPU (\d+)/KVM\n$`) // These threads follow this naming pattern as their command value (/proc/{pid}/task/{taskid}/comm)
-	// QEMU uses threads to represent vCPUs.
-}
-
-// Implement ShouldRunPrivileged method for QemuHypervisor
-func (q *QemuHypervisor) ShouldRunPrivileged() bool {
-	return false
-}
-
-// Implement GetHypervisorDevice method for QemuHypervisor
-func (q *QemuHypervisor) GetHypervisorDevice() string {
-	return "devices.kubevirt.io/kvm"
-}
-
-// Implement GetVirtLauncherMonitorOverhead method for QemuHypervisor
-func (q *QemuHypervisor) GetVirtLauncherMonitorOverhead() string {
-	return "25Mi"
-}
-
-// Implement GetVirtLauncherOverhead method for QemuHypervisor
-func (q *QemuHypervisor) GetVirtLauncherOverhead() string {
-	return "100Mi"
-}
-
-// Implement GetVirtlogdOverhead method for QemuHypervisor
-func (q *QemuHypervisor) GetVirtlogdOverhead() string {
-	return "20Mi"
-}
-
-// Implement GetHypervisorDaemonOverhead method for QemuHypervisor
-func (q *QemuHypervisor) GetHypervisorDaemonOverhead() string {
-	return "35Mi"
-}
-
-// Implement GetHypervisorOverhead method for QemuHypervisor
-func (q *QemuHypervisor) GetHypervisorOverhead() string {
-	return "30Mi"
-}
-
-// Implement GetDefaultKernelPath method for QemuHypervisor
-func (q *QemuHypervisor) GetDefaultKernelPath() (string, string) {
-	return "", ""
-}
-
-// Implement SupportsMemoryBallooning method for CloudHypervisor
-func (c *CloudHypervisor) SupportsMemoryBallooning() bool {
-	return false
-}
-
-// Implement RequiresBoot order method for CloudHypervisor
-func (c *CloudHypervisor) RequiresBootOrder() bool {
-	return true
-}
-
-// Implement GetDiskDriver method for CloudHypervisor
-func (c *CloudHypervisor) GetDiskDriver() string {
-	return "raw"
-}
-
-// Implement GetLibvirtSocketPath method for CloudHypervisor
-func (c *CloudHypervisor) GetLibvirtSocketPath() string {
-	return "libvirt/ch-sock" // TODO: Check this
-}
-
-// Implement GetVcpuRegex method for CloudHypervisor
-func (c *CloudHypervisor) GetVcpuRegex() *regexp.Regexp {
-	// parse thread comm value expression for MSHV
-	return regexp.MustCompile(`^vcpu(\d+)\n$`) // These threads follow this naming pattern as their command value (/proc/{pid}/task/{taskid}/comm)
-}
-
-// Implement ShouldRunPrivileged method for CloudHypervisor
-func (c *CloudHypervisor) ShouldRunPrivileged() bool {
-	return true
-}
-
-// Implement GetHypervisorDevice method for CloudHypervisor
-func (c *CloudHypervisor) GetHypervisorDevice() string {
-	return "devices.kubevirt.io/mshv"
-}
-
-// Implement GetVirtLauncherMonitorOverhead method for CloudHypervisor
-func (c *CloudHypervisor) GetVirtLauncherMonitorOverhead() string {
-	return "25Mi"
-}
-
-// Implement GetVirtLauncherOverhead method for CloudHypervisor
-func (c *CloudHypervisor) GetVirtLauncherOverhead() string {
-	return "100Mi"
-}
-
-// Implement GetVirtlogdOverhead method for CloudHypervisor
-func (c *CloudHypervisor) GetVirtlogdOverhead() string {
-	return "20Mi"
-}
-
-// Implement GetHypervisorDaemonOverhead method for CloudHypervisor
-func (c *CloudHypervisor) GetHypervisorDaemonOverhead() string {
-	return "35Mi"
-}
-
-// Implement GetHypervisorOverhead method for CloudHypervisor
-func (c *CloudHypervisor) GetHypervisorOverhead() string {
-	return "30Mi"
-}
-
-// Implement GetDefaultKernelPath method for CloudHypervisor
-func (c *CloudHypervisor) GetDefaultKernelPath() (string, string) {
-	return "/usr/share/cloud-hypervisor/CLOUDHV_EFI.fd", ""
-}
-
-// Implement GetDomainType method for CloudHypervisor
-func (c *CloudHypervisor) GetDomainType() string {
-	return "hyperv"
 }
 
 func NewHypervisor(hypervisor string) Hypervisor {
